+++ library/Encryption/Utils.hs
@@ -1,6 +1,6 @@
 module Encryption.Utils where
 
-import qualified Data.ByteString               as B
+import qualified Data.ByteString.Lazy          as B
 import qualified Data.Bits
 import           Numeric                        ( showHex )
 import           Data.Char                      ( toUpper )
@@ -11,13 +11,16 @@ snoc lst x = lst ++ [x]
 splitEvery :: Int -> B.ByteString -> [B.ByteString]
 splitEvery n list | list == B.empty = []
                   | otherwise       = first : splitEvery n rest
-  where (first, rest) = B.splitAt n list
+  where (first, rest) = B.splitAt (fromIntegral n) list
 
 rotWordLeft :: Int -> B.ByteString -> B.ByteString
-rotWordLeft n word = B.drop n word `B.append` B.take n word
+rotWordLeft n word =
+  B.drop (fromIntegral n) word `B.append` B.take (fromIntegral n) word
 
 rotWordRight :: Int -> B.ByteString -> B.ByteString
-rotWordRight n word = B.drop (len - n) word `B.append` B.take (len - n) word
+rotWordRight n word =
+  B.drop (len - fromIntegral n) word
+    `B.append` B.take (len - fromIntegral n) word
   where len = B.length word
 
 bsXor :: B.ByteString -> B.ByteString -> B.ByteString
@@ -40,4 +43,6 @@ zfill :: Int -> String -> String
 zfill num str = replicate (num - length str) '0' ++ str
 
 -- genCounts :: InitializationVector -> [InitializationVector]
+
 -- genCounts nonce = []
+
